requirements.txt
Django==1.8.6

djangorestframework==3.3.1
djangorestframework-jwt==1.7.2
django-rest-swagger==0.3.4

django-cors-headers==1.1.0
django-extensions==1.5.9
django-filter==0.10.0

factory_boy==2.6.0

markdown==2.6.2
werkzeug==0.10.1
django-debug-toolbar==1.4
coverage==4.0.2
Pillow==3.0.0
ipython==4.0.0
ipdb==0.8.1
mysqlclient==1.3.7
